+++ tests/BlazorServerFunctions.Generator.UnitTests/_snapshots/GrpcServerGeneratorTests/GrpcServerGeneratorTests.Generate_BasicGrpcInterface_ProducesServiceClass#ServerFunctionEndpointsRegistration.g.verified.cs
@@ -1,0 +1,1 @@
+﻿//HintName: ServerFunctionEndpointsRegistration.g.cs